Add ability to filter widgets by regex
Server side does not allow to filter by regex, so this is done on
client side. First the request is made with all the plain filters
(if they are specified, otherwise request the whole widgets json)
and then the received body could be filtered by regex.

Related ticket: https://progress.opensuse.org/issues/66769