Rename selected_item method to value for combobox
As we also can have editable comboboxes, `value` does reflect the
meaning of the method better, as we rely on the `value` property and not
traversing the list of items to find one where property `selected` is
set to true. Those values might conflict, so we should better rely on
the `value` property.