Introduce and demostrate configuration for ci release
package folder contains spec file for rpm build. This is the minimal spec that builds but raises 2 errors that we need to address.
Some tests moved to the test folder which is used in the ci. the names of the tests should end in _spec.rb or _test.rb. it does not matter which one it is used. You can run it with `rake test:unit`. The test in the widget folder needs to be adjusted.
The .travis will run the docker container and it will run tests and checks and then will build the rpm.

https://progress.opensuse.org/issues/70504